NURBS

Documentation for NURBS.

NURBS.Bspline

NURBS.BsplineCurve

NURBS.BsplineSurface

NURBS.NURB

NURBS.NURBScurve

NURBS.NURBSsurface

NURBS.bSpline

NURBS.bSplineDerivatives

NURBS.bSplineNaive

NURBS.bSplineNaive

NURBS.bSplineNaive

NURBS.bSplineNaive

NURBS.bSplineNaiveDerivative

NURBS.bSplineNaiveDerivative

NURBS.bSplineNaiveDerivative

NURBS.basisFun

NURBS.curvePoints

NURBS.curvePoints

NURBS.curvePoints

NURBS.curvePoints

NURBS.derBasisFun

NURBS.findSpan

NURBS.generateKnotVec

NURBS.numBasisFunctions

NURBS.nurbsNaive

NURBS.nurbsNaive

NURBS.nurbsNaive

NURBS.parseCtrlPoints

NURBS.parseLine

NURBS.parseSinglePatch

NURBS.readMultipatch

NURBS.surfacePoints

NURBS.surfacePoints

NURBS.surfacePoints

NURBS.surfacePoints

NURBS.Bspline — Type

Bspline{F} <: Basis

B-spline basis.

NURBS.BsplineCurve — Type

BsplineCurve{F} <: Curve

B-spline curve defined by the basis and the control points.

NURBS.BsplineSurface — Type

BsplineSurface{F} <: Surface

Surface defined by a B-spline basis and the control points.

NURBS.NURB — Type

NURB{F} <: Basis

NURBS basis.

NURBS.NURBScurve — Type

NURBScurve{F} <: Curve

B-spline curve defined by the basis and the control points.

NURBS.NURBSsurface — Type

NURBSsurface{F} <: Surface

Surface defined by a B-spline basis, the control points, and the weights.

NURBS.bSpline — Method

bSpline(basis::Bspline, evalpoints)

Evaluate B-spline basis at all evalpoints.

NURBS.bSplineDerivatives — Method

bSplineDerivatives(basis::Bspline, k::Int, evalpoints)

Evaluate k-the derivative of B-spline basis at all evalpoints (all basis functions different from 0 at the evalpoints are evaluated).

NURBS.bSplineNaive — Method

bSplineNaive(knotVector, i::Int, degree::Int, evalpoints; normalize=true)

i-th b-spline basis function of degree 'degree' evaluated at all 'evalpoints'.

The knotvector is assumed to be normalized to [1, 0].

NURBS.bSplineNaive — Method

bSplineNaive(knotVector, i::Int, degree::Int, u::Real)

i-th b-spline basis function of degree 'degree' evaluated at u.

Formula (2.5) of 'The NURBS Book' p. 50.

NURBS.bSplineNaive — Method

bSplineNaive(knotVector, i::Int, u::Real)

i-th b-spline basis function of degree 0 evaluated at u.

Formula (2.5) of 'The NURBS Book' p. 50.

NURBS.bSplineNaive — Method

bSplineNaive(basis::Bspline, i::Int, evalpoints)

i-th B-spline basis function evaluated at all 'evalpoints'.

NURBS.bSplineNaiveDerivative — Method

bSplineNaiveDer(knotVector, i::Int, degree::Int, evalpoints, k::Int; normalize=true)

Compute the k-th derivative of i-th b-spline basis function of degree 'degree' evaluated at all 'evalpoints'.

NURBS.bSplineNaiveDerivative — Method

bSplineNaiveDerivative(knotVector, i::Int, degree::Int, u::Real, k::Int)

k-th derivative of the i-th b-spline basis function of degree 'degree' evaluated at (single) point 'u'.

Formula (2.9) of 'The NURBS Book' p. 61. (Recursive implementation to avoid the faculties in the (2.10) formula.)

NURBS.bSplineNaiveDerivative — Method

bSplineNaiveDerivative(basis::Bspline, i::Int, k::Int, evalpoints)

Compute the k-th derivative of i-th b-spline basis function evaluated at all 'evalpoints'.

NURBS.basisFun — Method

basisFun(knotSpan, uVector, degree::Int, knotVector)

Compute the nonvanishing B-spline basis functions of degree 'degree' at the parametric points defined by 'uVector'

Return the B-spline basis functions vector of size length(uVector) * (degree + 1).

Adapted from Algorithm A2.2 from 'The NURBS Book' p. 70.

NURBS.curvePoints — Method

curvePoints(curve::BsplineCurve, uVector)

Convenience function to plot a NURBS curve.

NURBS.curvePoints — Method

curvePoints(nbasisFun::Int, degree::Int, knotVector, controlPoints, uVector, weights)

Compute a 1D NURBS curve: given the 'knotVector', the 'controlPoints', the 'degree', and the 'weights', the curve is evaluated at the points given in 'uVector'.

Example for the controlPoints:

P1 = SVector(0.0, 0.0, 0.0) P2 = SVector(0.1, 0.25, 0.0) P3 = SVector(0.25, 0.3, 0.0)

controlPoints = [P1, P2, P3]

NURBS.curvePoints — Method

curvePoints(nbasisFun::Int, degree::Int, knotVector, controlPoints, uVector)

Compute a 1D B-spline curve: given the 'knotVector', the 'controlPoints', and the 'degree', the curve is evaluated at the points given in 'uVector'.

Example for the controlPoints:

P1 = SVector(0.0, 0.0, 0.0) P2 = SVector(0.1, 0.25, 0.0) P3 = SVector(0.25, 0.3, 0.0)

controlPoints = [P1, P2, P3]

NURBS.curvePoints — Method

curvePoints(curve::NURBScurve, uVector)

Convenience function to plot a NURBS curve.

NURBS.derBasisFun — Method

derBasisFun(knotSpan, degree::Int, evalpoints, knotVector, numberDerivatives::Int)

Compute the nonvanishing B-spline basis functions and its derivatives of degree 'degree' at the parametric points defined by 'uVector'.

Organization of output: dersv[n, k, :] contains (k-1)-th derivative at n-th point.

Adapted from Algorithm A2.3 from 'The NURBS BOOK' p. 72.

NURBS.findSpan — Method

findSpan(n::Int, u, knotVector)

Find the spans of a B-spline knot vector at the parametric points 'u', where 'b' is the number of basis functions (control points).

Span: the intervall index in which a point lies. E.g., knotVector = [0, 1, 2, 3, 4]. Hence, there are 4 intervalls. u=1.2 lies in the second intervall.

Modification of Algorithm A2.1 from 'The NURBS Book' p. 68.

Assumption that the knotVector is open! (the first and last knot are repeated degree + 1 times)

NURBS.generateKnotVec — Method

generateKnotVec(b::Int, degree::Int)

Convenience function to generate a knot vector for 'b' basis functions and a certain 'degree':

The first and last entry are repeated 'degree'+1 times. Normalized to [0, 1].

NURBS.numBasisFunctions — Method

numBasisFunctions(basis::Basis)

The number of basis functions is fixed by the knot vector and the degree.

Assumption: the first and last knot vector entry has mulitplicity degree + 1.

NURBS.nurbsNaive — Method

nurbsNaive(knotVector, i::Int, degree::Int, evalpoints, weights)

i-th NURB basis function of degree 'degree', the 'knotVector', and with 'weights' evaluated at all 'evalpoints'.

NURBS.nurbsNaive — Method

nurbsNaive(knotVector, i::Int, degree::Int, u::Real, weights)

i-th NURB basis function of degree 'degree', the 'knotVector', and with 'weights' evaluated at single point 'u'.

Formula (4.2) of 'The NURBS Book' p. 118.

NURBS.nurbsNaive — Method

nurbsNaive(basis::NURB, i::Int, evalpoints)

i-th NURB basis function evaluated at all 'evalpoints'.

NURBS.parseCtrlPoints — Function

parseCtrlPoints(pointDims, stringVec, T=Float64)

Take the 3 strings in 'stringVec' and parse it into a pointDims[1] x pointDims[2] matrix where each entry is a SVector for a controlpoint.

NURBS.parseLine — Function

parseLine(line::String, T=Float64)

Take the string 'line' of the form '0.1 0.2 0.6 ...' and parse it to a vector with eltype 'T'.

NURBS.parseSinglePatch — Method

parseSinglePatch(stringVec, lineInd)

Take the string vector and the line index where "PATCH" stands and extract the information of a single patch

NURBS.readMultipatch — Function

readMultipatch(filename::String)

Read multipatch file organized as:

PATCH a 4 4 -> degree in u and v 5 5 -> number of control points in u and v 0.0 0.0 ... -> knot vector in u 0.0 0.0 ... -> knot vector in v 0.1 0.2 ... | 0.1 0.2 ... > xyz components of the control points (normalized with the weigths -> we remove this weighting when reading in the data) 0.1 0.2 ... | 1.0 1.0 ... -> weights

PATCH b ...

NURBS.surfacePoints — Method

surfacePoints(Patch::BsplineSurface, uEvalpoints, vEvalpoints)

Convenience function to plot a B-spline surface.

NURBS.surfacePoints — Method

surfacePoints(uDegree::Int, vDegree::Int, uKnotVector, vKnotVector, controlPoints, uVector, vVector, weights)

Compute NURBS surface: given the knotvectors and the degrees in 'u' and 'v' direction, the surface is evaluated at the evaluation points (uVector, vVector).

Control points ordering P_(xi,yj):

P11 ––- P12 ––- P13 –-> y / v direction | | | | | | P21 ––- P22 ––- P23 | | | | | | P31 ––- P32 ––- P_33 | x / u direction

NURBS.surfacePoints — Method

surfacePoints(uDegree::Int, vDegree::Int, uKnotVector, vKnotVector, controlPoints, uVector, vVector)

Compute B-spline surface: given the knotvectors and the degrees in 'u' and 'v' direction, the surface is evaluated at the evaluation points (uVector, vVector).

Control points ordering P_(xi,yj):

P11 ––- P12 ––- P13 –-> y / v direction | | | | | | P21 ––- P22 ––- P23 | | | | | | P31 ––- P32 ––- P_33 | x / u direction

NURBS.surfacePoints — Method

surfacePoints(Patch::NURBSsurface, uEvalpoints, vEvalpoints)

Convenience function to plot a NURBSsurface.